requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.2.0
attrs==25.4.0
babel==2.18.0
beautifulsoup4==4.14.3
bigtree==1.3.1
bleach==6.3.0
blinker==1.9.0
cachetools==7.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
dask==2026.1.2
databricks-sdk==0.91.0
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.9
distro==1.9.0
docker==7.1.0
dynaconf==3.2.12
evidently==0.7.20
executing==2.2.1
Faker==40.4.0
fastapi==0.129.0
fastjsonschema==2.21.2
feast==0.61.0
filelock==3.24.3
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.61.1
fqdn==1.5.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.48.0
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
greenlet==3.3.2
gunicorn==25.1.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huey==2.6.0
idna==3.11
imbalanced-learn==0.14.1
importlib_metadata==8.7.1
iniconfig==2.3.0
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
iterative-telemetry==0.0.10
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
json5==0.13.0
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kafka-python==2.3.0
kiwisolver==1.4.9
lark==1.3.1
librt==0.8.1
litestar==2.21.0
litestar-htmx==0.5.0
locket==1.0.0
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mlflow==3.10.0
mlflow-skinny==3.10.0
mlflow-tracing==3.10.0
mmh3==5.2.1
msgspec==0.20.0
multidict==6.7.1
multipart==1.3.0
mypy==1.19.1
mypy_extensions==1.1.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.9.2
notebook==7.5.3
notebook_shim==0.2.4
numpy==2.4.2
opentelemetry-api==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
packaging==26.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.6
partd==1.4.2
pathspec==1.0.4
patsy==1.0.2
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
plotly==5.24.1
pluggy==1.6.0
polyfactory==3.2.0
prettytable==3.17.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rich-click==1.9.7
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
Send2Trash==2.1.0
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.5
skops==0.13.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.52.1
statsmodels==0.14.6
tabulate==0.10.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typeguard==4.5.1
typer==0.24.0
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ujson==5.11.0
uri-template==1.3.0
urllib3==2.6.3
uuid6==2025.0.1
uvicorn==0.34.0
uvicorn-worker==0.3.0
uvloop==0.22.1
watchdog==6.0.0
watchfiles==1.1.1
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
Werkzeug==3.1.6
widgetsnbextension==4.0.15
xgboost==3.2.0
zipp==3.23.0
